Refactor : 엔티티 기본키 Long -> Integer

See merge request s10-webmobile2-sub2/S10P12A610!29

+++ backend/src/main/java/com/ssafy/togeball/domain/chatroom/repository/ChatroomMembershipRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ChatroomMembershipRepository extends JpaRepository<ChatroomMembership, Long> {
+public interface ChatroomMembershipRepository extends JpaRepository<ChatroomMembership, Integer> {
 
-    Optional<ChatroomMembership> findByUserIdAndChatroomId(Long userId, Long chatroomId);
+    Optional<ChatroomMembership> findByUserIdAndChatroomId(Integer userId, Integer chatroomId);
 }
